# openapi.yml
openapi: 3.0.3
info:
  title: Space Marine Management API
  version: 1.0.0
  description: API for managing Space Marines, their Coordinates, and Chapters with full CRUD, pagination support, special operations, embedded entities, and export capabilities.
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
...
  /space-marines:
    get:
      summary: Get all Space Marines (paginated)
      operationId: getAllSpaceMarines
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
        - name: embed
          in: query
          required: false
          schema:
            type: string
            enum: [all, coordinates, chapter]
          description: |
            Embed related entities. Supported values:
            - `all`: Embed both coordinates and chapter
            - `coordinates`: Embed only coordinates
            - `chapter`: Embed only chapter
          example: all
      responses:
        "200":
          description: |
            Paginated list of Space Marines. When `embed` parameter is used,
            response contains embedded entities without IDs.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PaginatedSpaceMarineResponse"
                  - $ref: "#/components/schemas/PaginatedSpaceMarineEmbeddedResponse"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new Space Marine
      operationId: createSpaceMarine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceMarineCreateRequest"
      responses:
        "201":
          description: Space Marine created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceMarine"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Referenced Coordinates or Chapter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /space-marines/{id}:
    get:
      summary: Get a Space Marine by ID
      operationId: getSpaceMarineById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: embed
          in: query
          required: false
          schema:
            type: string
            enum: [all, coordinates, chapter]
          description: |
            Embed related entities. Supported values:
            - `all`: Embed both coordinates and chapter
            - `coordinates`: Embed only coordinates
            - `chapter`: Embed only chapter
          example: all
      responses:
        "200":
          description: |
            Space Marine found. When `embed` parameter is used,
            response contains embedded entities without IDs.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SpaceMarine"
                  - $ref: "#/components/schemas/SpaceMarineEmbedded"
        "404":
          description: Space Marine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update a Space Marine by ID
      operationId: updateSpaceMarine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceMarineUpdateRequest"
      responses:
        "200":
          description: Space Marine updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceMarine"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Space Marine or referenced entities not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a Space Marine by ID
      operationId: deleteSpaceMarine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        "204":
          description: Space Marine deleted
        "404":
          description: Space Marine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /space-marines/export/json:
    get:
      summary: Export Space Marines as JSON
      operationId: exportSpaceMarinesJson
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
            default: 1000
          description: Number of items per page (max 1000 for exports)
        - name: embed
          in: query
          required: false
          schema:
            type: string
            enum: [all, coordinates, chapter]
          description: |
            Embed related entities in exported data. Supported values:
            - `all`: Embed both coordinates and chapter
            - `coordinates`: Embed only coordinates
            - `chapter`: Embed only chapter
      responses:
        "200":
          description: JSON file download with paginated Space Marines data
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: "#/components/schemas/SpaceMarine"
                    - $ref: "#/components/schemas/SpaceMarineEmbedded"
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "space_marines_page_0_size_1000.json"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /space-marines/export/xml:
    get:
      summary: Export Space Marines as XML
      operationId: exportSpaceMarinesXml
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
            default: 1000
          description: Number of items per page (max 1000 for exports)
        - name: embed
          in: query
          required: false
          schema:
            type: string
            enum: [all, coordinates, chapter]
          description: |
            Embed related entities in exported data. Supported values:
            - `all`: Embed both coordinates and chapter
            - `coordinates`: Embed only coordinates
            - `chapter`: Embed only chapter
      responses:
        "200":
          description: XML file download with paginated Space Marines data
          content:
            application/xml:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: "#/components/schemas/SpaceMarine"
                    - $ref: "#/components/schemas/SpaceMarineEmbedded"
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "space_marines_page_0_size_1000.xml"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ============
  # COORDINATES
  # ============
...
components:
  schemas:
...
    PaginatedSpaceMarineResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/SpaceMarine"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
...
    PaginatedSpaceMarineEmbeddedResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/SpaceMarineEmbedded"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0

    # === SPACE MARINE ===
    SpaceMarine:
      type: object
      required:
        - id
        - name
        - coordinatesId
        - creationDate
        - chapterId
        - health
        - weaponType
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          description: Unique auto-generated ID (> 0)
        name:
          type: string
          minLength: 1
          description: Non-empty, non-null name
        coordinatesId:
          type: integer
          format: int64
          minimum: 1
          description: ID of associated Coordinates
        creationDate:
          type: string
          format: date-time
          description: Auto-generated creation timestamp
        chapterId:
          type: integer
          format: int64
          minimum: 1
          description: ID of associated Chapter
        health:
          type: integer
          format: int64
          minimum: 1
          description: Health must be > 0
        loyal:
          type: boolean
          nullable: true
        category:
          type: string
          nullable: true
          enum: [AGGRESSOR, INCEPTOR, TACTICAL, CHAPLAIN, APOTHECARY]
          description: Astartes category
        weaponType:
          type: string
          enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]
          description: Type of weapon used by the Space Marine

    # === EMBEDDED SPACE MARINE ===
    SpaceMarineEmbedded:
      type: object
      required:
        - id
        - name
        - creationDate
        - health
        - weaponType
        - coordinates
        - chapter
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          description: Unique auto-generated ID (> 0)
        name:
          type: string
          minLength: 1
          description: Non-empty, non-null name
        creationDate:
          type: string
          format: date-time
          description: Auto-generated creation timestamp
        health:
          type: integer
          format: int64
          minimum: 1
          description: Health must be > 0
        loyal:
          type: boolean
          nullable: true
        category:
          type: string
          nullable: true
          enum: [AGGRESSOR, INCEPTOR, TACTICAL, CHAPLAIN, APOTHECARY]
          description: Astartes category
        weaponType:
          type: string
          enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]
          description: Type of weapon used by the Space Marine
        coordinates:
          $ref: "#/components/schemas/Coordinates"
          description: Embedded coordinates data without ID
        chapter:
          $ref: "#/components/schemas/Chapter"
          description: Embedded chapter data without ID
    SpaceMarineCreateRequest:
      type: object
      required:
        - name
        - coordinatesId
        - chapterId
        - health
        - weaponType
      properties:
        name:
          type: string
          minLength: 1
        coordinatesId:
          type: integer
          format: int64
          minimum: 1
        chapterId:
          type: integer
          format: int64
          minimum: 1
        health:
          type: integer
          format: int64
          minimum: 1
        loyal:
          type: boolean
          nullable: true
        category:
          type: string
          nullable: true
          enum: [AGGRESSOR, INCEPTOR, TACTICAL, CHAPLAIN, APOTHECARY]
        weaponType:
          type: string
          enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]
    SpaceMarineUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        coordinatesId:
          type: integer
          format: int64
          minimum: 1
        chapterId:
          type: integer
          format: int64
          minimum: 1
        health:
          type: integer
          format: int64
          minimum: 1
        loyal:
          type: boolean
          nullable: true
        category:
          type: string
          nullable: true
          enum: [AGGRESSOR, INCEPTOR, TACTICAL, CHAPLAIN, APOTHECARY]
        weaponType:
          type: string
          enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]

    # === COORDINATES ===
    Coordinates:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: integer
          format: int64
          description: X coordinate value
        y:
          type: number
          format: float
          maximum: 343
          description: Y coordinate value (must be ≤ 343)
...
    Chapter:
      type: object
      required:
        - name
        - marinesCount
      properties:
        name:
          type: string
          minLength: 1
          description: Name of the Chapter
        marinesCount:
          type: integer
          format: int64
          minimum: 1
          maximum: 1000
          description: Number of marines in the chapter (must be between 1 and 1000)
...
    ErrorResponse:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
      example:
        error: "Invalid weapon types: LASER_CANNON. Valid values: BOLTGUN,HEAVY_BOLTGUN,FLAMER,HEAVY_FLAMER,MULTI_MELTA"
        timestamp: "2025-11-07T14:30:00.000Z"